CHANGES REQUESTED IMPLEMENTED
Changes requested for checkout and sign up has been implemented.
